docs: initial documentation for reworked ODE classes
